09/03/2020

+++ app/Http/Controllers/Backend/UserController.php
@@ -4,8 +4,11 @@ namespace App\Http\Controllers\Backend;
 
 use App\Http\Controllers\Controller;
 use App\Models\Book;
+use App\Models\Shop;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Session;
 use function Ramsey\Uuid\v1;
 
 class UserController extends Controller
@@ -33,7 +36,13 @@ class UserController extends Controller
      */
     public function create()
     {
-        return view('backend.users.create');
+        $user=Auth::user();
+        $this->authorize('create',User::class);
+
+//        $this->authorize('create', $user);
+        return view('backend.users.create')->with([
+            'user'=>Auth::user()
+        ]);
     }
 
     /**
@@ -44,7 +53,28 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        dd($request);
+        $user=new User();
+        $user->name=$request->name;
+        $user->email=$request->email;
+        $user->phone=$request->phone;
+        $user->address=$request->address;
+//        $user->status=1;
+        $user->role=1;
+        $avatar=$request->file('avatar');
+        $avatar->store('image');
+
+        $name = date('YmdHis') ."." . $avatar->getClientOriginalExtension();
+        $avatar->move('backend/dist/img',$name);
+        $user->avatar=$name;
+        $user->password = md5($request->password);
+
+        $user->save();
+//        dd($user->name);
+        if(Auth::user()->role=2) {
+            return redirect()->route('backend.user.show', $user->id);
+        }else{
+            return redirect()->route('frontend.home.showAccount',$user->id);
+        }
     }
 
     /**
@@ -56,8 +86,22 @@ class UserController extends Controller
     public function show($id)
     {
         $user=User::find($id);
+        $shop=Shop::where('user_id',$id)->first();
+//        dd($shop);
         return view('backend.users.show')->with([
-            'user'=>$user
+            'userr'=>$user,
+            'shop'=>$shop,
+            'user'=>Auth::user()
+        ]);
+    }
+    public function showShop($id)
+    {
+        $shop=Shop::find($id);
+        $user=User::find($shop->user_id);
+        return view('backend.users.show')->with([
+            'userr'=>$user,
+            'shop'=>$shop,
+            'user'=>Auth::user()
         ]);
     }
 
@@ -70,6 +114,7 @@ class UserController extends Controller
     public function edit($id)
     {
         $user=User::find($id);
+        $this->authorize('edit',$user);
         //nếu người dùng role=1 thì trả về giao diện khác
         if($user->role==1){
             return view('frontend.page.editAccount')->with([
@@ -104,13 +149,21 @@ class UserController extends Controller
         $user->email=$request->email;
         $user->phone=$request->phone;
         $user->address=$request->address;
-        $user->role=$request->role;
+        $avatar=$request->file('avatar');
+        $avatar->store('image');
+        $name = date('YmdHis') ."." . $avatar->getClientOriginalExtension();
+        $avatar->move('backend/dist/img',$name);
+        $user->avatar=$name;
         if($request->password !='null') {
             $user->password = md5($request->password);
         }
         $user->save();
 //        dd($user->name);
-        return redirect('/');
+        if($user->role=2) {
+            return redirect()->route('backend.user.show', $user->id);
+        }else{
+            return redirect()->route('frontend.home.showAccount',$user->id);
+        }
     }
 
     /**
@@ -122,9 +175,87 @@ class UserController extends Controller
     public function destroy($id)
     {
         $user=User::find($id);
+        $this->authorize('delete',$user);
 //        dd($user);
         $user->delete();
         return redirect('/admin/users');
 
+    }
+    public function formUpgrade($id){
+        return view('frontend.page.upgrade')->with([
+            'user'=>Auth::user()
+        ]);
+    }
+    public function storeShop(Request $request){
+//        dd("aaaaaaaaaaaaaaaa");
+//        dd($request);
+        $shop=new Shop();
+        $shop->name=$request->name;
+        $shop->phone=$request->phone;
+        $shop->country=$request->country;
+        $shop->address=$request->address;
+        $shop->supplier=$request->supplier;
+        $shop->status=0;
+        if($request->policy) {
+            $shop->check_policy = 1;
+        }else{
+            return redirect()->back();
+        }
+        $shop->user_id=Auth::user()->id;
+        $shop->save();
+//        dd($shop);
+        Session::put('waite','Yêu cầu của bạn đang đợi quản trị viên phê duyệt',5);
+        return redirect('/home');
+
+    }
+    public function listUpgrade(){
+//        $this->authorize('delete',User::class);
+
+        $list=Shop::where('status','0')->get();
+        return view('backend.users.listUpgrade')->with([
+            'list'=>$list,
+            'user'=>Auth::user()
+        ]);
+    }
+    public function listShop(){
+        $list=Shop::where('status','1')->get();
+        return view('backend.users.listShop')->with([
+            'list'=>$list,
+            'user'=>Auth::user()
+        ]);
+    }
+    public function listDeletedShop(){
+        $user=Auth::user();
+        $this->authorize('delete',User::class);
+
+        $list=Shop::where('status','2')->get();
+        return view('backend.users.listDeletedShop')->with([
+            'list'=>$list,
+            'user'=>Auth::user()
+        ]);
+    }
+
+    public function upgrade($id){
+//        $this->authorize('delete', User::find($id));
+
+        $shop=Shop::find($id);
+        $shop->status=1;
+        $user=User::find($shop->user_id);
+//        $this->authorize('delete',$user);
+
+        $user->role=2;
+        $user->save();
+        $shop->save();
+        return redirect()->back();
+    }
+    public function destroyShop($id){
+        $this->authorize('delete', User::find($id));
+        $shop=Shop::find($id);
+        $shop->status=2;
+        $user=User::find($shop->user_id);
+        $user->role=1;
+        $user->save();
+        $shop->save();
+        return redirect()->back();
     }
 }
